pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "hypha-agent-engine"
version = "0.1.0"
authors = [
    {name = "Wei Ouyang", email = "oeway007@gmail.com"},
]
description = "Hypha Agent Platform Backend Engine for LLM Services and Tool Runtime"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv",
    "schema-agents==0.2.0rc2",
    "hypha-rpc>=0.20.48",
    "requests",
    "pypdf",
    "pillow",
    "matplotlib",
    "hypha>=0.20.47",
    "minio>=7.2.8",
    "tqdm",
    "aiofiles",
    "langchain>=0.2.11",
    "beautifulsoup4",
    "pandas",
    "duckduckgo-search>=6.1.5",
    "rank-bm25",
    "langchain-openai>=0.1.19",
    "langchain-core>=0.2.24",
    "langchain-community>=0.2.10",
    "html2text",
    "fastapi",
    "asyncpg",
    "greenlet",
]

[tool.poetry]
name = "hypha-agent-engine"
version = "0.1.0"
description = "Backend service for AI agent runtime"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
hypha-rpc = "*"
schema-agents = "*"
click = "*"
pydantic = "*"

[tool.poetry.scripts]
hypha-agent-engine = "hypha_agent_engine.__main__:cli"
